Web Scraper Course Project/requirements.txt
beautifulsoup4==4.9.3
requests==2.27.1
Pillow~=6.2.2
argparse~=1.2.1
mock~=3.0.5
